requirements.txt
openai
langchain
huggingface_hub
python-decouple